pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lore"
version = "0.1.0"
description = "Local semantic knowledge management for developers"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "anthropic>=0.40",
    "python-dotenv>=1.0",
]

[project.scripts]
lore = "lore.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
